Added indication for closed issues in notes
For issues that are closed, the links will now show '[closed]' following
the issue number. This is done as post-process after the markdown has
been loaded from the cache as the status of the issue may change between
the cache being populated and the note being displayed.

Also as part of this change, the Banzai
BaseParser#grouped_objects_for_nodes method has been refactored to return a
Hash utilising the node itself as the key, since this was a common
pattern of usage for this method.