Fix Projects::UploadService spec
For reasons unknown, this spec was passing a project's repository as the
first argument to `Projects::UploadService`, while it was expecting the
project itself.

Previously this wasn't a problem because `FileUploader` only needed the
object to respond to `path_with_namespace`, which Repository and Project
both did. But now because of `Upload` and its polymorphic association,
it expects the object to respond to `primary_key`.